Add a CoverCalibratorAdapter, and some related enums and enumerator values. TODO Add match_literals.* support for CoverCalibrator.
Move device_impl_base.* into src/device_types/, and move adapters into
device type specific subfolders of that new folder.

Update README.md with a bunch of notes about generating adapters
and string literals in PROGMEM.

+++ examples/TinyAlpacaServerDemo/dht22_handler.cpp
@@ -39,11 +39,11 @@ const alpaca::DeviceInfo kDht22DeviceInfo{
 
 Dht22Handler::Dht22Handler() : ObservingConditionsAdapter(kDht22DeviceInfo) {}
 
-StatusOr<float> Dht22Handler::GetHumidity() {
+StatusOr<double> Dht22Handler::GetHumidity() {
   return dht22.get_relative_humidity();
 }
 
-StatusOr<float> Dht22Handler::GetTemperature() {
+StatusOr<double> Dht22Handler::GetTemperature() {
   return dht22.get_temperature();
 }
